refactor(cli): standardize test organization to colocation pattern
- Remove duplicate tests between tests/ and src/ directories
- Migrate all tests to src/*/__tests__/ colocation structure
- Organize special test types (performance, integration) in subdirectories
- Update vitest configuration to remove tests/ pattern
- Fix import paths for moved test utilities
- Remove empty tests/ directory structure

Benefits:
- Eliminates duplicate test coverage
- Follows 2025 best practices for test colocation
- Improves maintainability with tests next to source code
- Simplifies configuration with single test pattern
- Better IDE performance and test discovery
