+++ packages/cli/src/core/validation/__tests__/domain.test.ts
