+++ packages/cli/src/core/validation/__tests__/pipeline.test.ts
