+++ packages/cli/src/testing/test-utils/cross-platform-paths.ts
